+++ lib/sh/libuiDate.sh
@@ -27,7 +27,7 @@
 #
 #####
 
-Version -r 1.829 -m 1.5
+Version -r 2.000 -m 1.5
 
 # defaults