+++ lib/sh/libuiPackage.sh
@@ -27,7 +27,7 @@
 #
 #####
 
-Version -r 1.835 -m 1.17
+Version -r 2.000 -m 1.17
 
 # defaults